[feat] #34 Store 관련 모델 Domain 모듈로 이동
