+++ data/src/main/java/org/prography/cakk/data/api/dto/response/StoreDetailResponse.kt
@@ -1,6 +1,7 @@
-package org.prography.cakk.data.api.model.response
+package org.prography.cakk.data.api.dto.response
 
 import kotlinx.serialization.Serializable
+import org.prography.domain.model.store.StoreDetailModel
 
 @Serializable
 data class StoreDetailResponse(
@@ -18,4 +19,21 @@ data class StoreDetailResponse(
     val name: String,
     val phoneNumber: String,
     val storeTypes: List<String>,
-)
+) {
+    fun toModel() = StoreDetailModel(
+        address = address,
+        city = city,
+        createdAt = createdAt,
+        description = description,
+        district = district,
+        id = id,
+        latitude = latitude,
+        link = link,
+        location = location,
+        longitude = longitude,
+        modifiedAt = modifiedAt,
+        name = name,
+        phoneNumber = phoneNumber,
+        storeTypes = storeTypes
+    )
+}
