+++ data/src/main/java/org/prography/cakk/data/api/dto/response/StoreListResponse.kt
@@ -1,6 +1,7 @@
-package org.prography.cakk.data.api.model.response
+package org.prography.cakk.data.api.dto.response
 
 import kotlinx.serialization.Serializable
+import org.prography.domain.model.store.StoreModel
 
 @Serializable
 data class StoreListResponse(
@@ -14,4 +15,17 @@ data class StoreListResponse(
     val latitude: Double,
     val longitude: Double,
     val storeTypes: List<String>,
-)
+) {
+    fun toModel() = StoreModel(
+        id = id,
+        createdAt = createdAt,
+        modifiedAt = modifiedAt,
+        name = name,
+        city = city,
+        district = district,
+        location = location,
+        latitude = latitude,
+        longitude = longitude,
+        storeTypes = storeTypes
+    )
+}
