+++ domain/src/main/java/org/prography/domain/model/store/StoreDetailModel.kt
@@ -1,0 +1,18 @@
+package org.prography.domain.model.store
+
+data class StoreDetailModel(
+    val address: String = "",
+    val city: String = "",
+    val createdAt: String = "",
+    val description: String = "",
+    val district: String = "",
+    val id: Int = 0,
+    val latitude: Double = 0.0,
+    val link: String = "",
+    val location: String = "",
+    val longitude: Double = 0.0,
+    val modifiedAt: String = "",
+    val name: String = "",
+    val phoneNumber: String = "",
+    val storeTypes: List<String> = listOf(),
+)
